pyproject.toml
[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[tool.poetry]
name = "bsms-int"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bsms_int"}]

[tool.poetry.dependencies]
python = "^3.10"
black = "^24.4.2"
numpy = "^2.0.1"
scipy = "^1.14.0"
torch = "^2.4.0"
sparse-dot-mkl = "^0.9.4"
torchdata = "^0.7.1"
hydra-core = "^1.3.2"
wandb = "^0.17.5"
pytz = "^2024.1"
matplotlib = "^3.9.1"
torchsummary = "^1.5.1"
h5py = "^3.11.0"
mkl = "^2024.2.0"
tabulate = "^0.9.0"
torch-geometric = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
